Add help functionality and map files
This commit adds a new file, help.c, which contains the implementation of the print_help function. This function prints a help message with usage instructions and available options. The commit also includes the addition of map files map_0, map_1, map_2, default, and new_map to the maps directory. These map files will be used to create the maze for the game.

The changes in this commit improve the user experience by providing helpful information and expanding the available map options.